Fix up compiler warnings emitted by VC7.1.

+++ asio/include/asio/detail/hash_map.hpp
@@ -28,9 +28,9 @@ namespace asio {
 namespace detail {
 
 template <typename K>
-unsigned int hash(const K& k)
+size_t hash(const K& k)
 {
-  return (unsigned int)k;
+  return (size_t)k;
 }
 
 template <typename K, typename V>
@@ -51,7 +51,7 @@ public:
   hash_map()
   {
     // Initialise all buckets to empty.
-    for (unsigned int i = 0; i < num_buckets; ++i)
+    for (size_t i = 0; i < num_buckets; ++i)
       buckets_[i].first = buckets_[i].last = values_.end();
   }
 
@@ -82,7 +82,7 @@ public:
   // Find an entry in the map.
   iterator find(const K& k)
   {
-    unsigned int bucket = hash(k) % num_buckets;
+    size_t bucket = hash(k) % num_buckets;
     iterator it = buckets_[bucket].first;
     if (it == values_.end())
       return values_.end();
@@ -100,7 +100,7 @@ public:
   // Find an entry in the map.
   const_iterator find(const K& k) const
   {
-    unsigned int bucket = hash(k) % num_buckets;
+    size_t bucket = hash(k) % num_buckets;
     const_iterator it = buckets_[bucket].first;
     if (it == values_.end())
       return it;
@@ -118,7 +118,7 @@ public:
   // Insert a new entry into the map.
   std::pair<iterator, bool> insert(const value_type& v)
   {
-    unsigned int bucket = hash(v.first) % num_buckets;
+    size_t bucket = hash(v.first) % num_buckets;
     iterator it = buckets_[bucket].first;
     if (it == values_.end())
     {
@@ -143,7 +143,7 @@ public:
   {
     assert(it != values_.end());
 
-    unsigned int bucket = hash(it->first) % num_buckets;
+    size_t bucket = hash(it->first) % num_buckets;
     bool is_first = (it == buckets_[bucket].first);
     bool is_last = (it == buckets_[bucket].last);
     if (is_first && is_last)
@@ -160,7 +160,7 @@ public:
   void clear()
   {
     // Initialise all buckets to empty.
-    for (unsigned int i = 0; i < num_buckets; ++i)
+    for (size_t i = 0; i < num_buckets; ++i)
       buckets_[i].first = buckets_[i].last = values_.end();
 
     // Clear the values.
